http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@508443 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/stomp/StompConnector.cpp
@@ -95,7 +95,7 @@ StompConnector::~StompConnector()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-unsigned int StompConnector::getNextProducerId()
+long long StompConnector::getNextProducerId()
 {
     synchronized( &mutex )
     {
@@ -106,7 +106,7 @@ unsigned int StompConnector::getNextProducerId()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-unsigned int StompConnector::getNextTransactionId()
+long long StompConnector::getNextTransactionId()
 {
     synchronized( &mutex )
     {
